requirements.txt
numpy
pandas
flask~=2.3.2
flask-login
plotly
dash
tabulate
secret
Pillow
kaleido
Flask-SocketIO